Implementacao do servico Inbound

+++ src/Services/InbounOrder/Inbound.Application/Queries/DTOs/OrderItemResponseDTO.cs
@@ -2,7 +2,7 @@
 {
     public class OrderItemResponseDTO
     {
-        public OrderItemResponseDTO(Guid id, string productCode, string typepackage, decimal quantityPackage, decimal quantity)
+        public OrderItemResponseDTO(Guid id, string productCode, string typepackage, int quantityPackage, decimal quantity)
         {
             Id = id;
             ProductCode = productCode;
@@ -14,7 +14,7 @@
         public Guid Id { get; private set; }
         public string ProductCode { get; private set; }
         public string TypePackage { get; private set; }
-        public decimal QuantityPackage { get; private set; }
+        public int QuantityPackage { get; private set; }
         public decimal Quantity { get; private set; }
     }
 }
